Auton almost done

+++ Parsec 2022 Worlds Code/include/subsystems/arm.hpp
@@ -24,7 +24,7 @@ class Arm{
 
 
         // ENCODER METHODS
-        int getEncoderRaw();
+        double getEncoderRaw();
 
         void resetEncoder();
