+++ Parsec 2022 Worlds Code/include/subsystems/stick.hpp
@@ -24,7 +24,7 @@ class Stick{
 
 
         // ENCODER METHODS
-        int getEncoderRaw();
+        double getEncoderRaw();
 
         void resetEncoder();
